Moved construct search into SearchApiMetadataSource to dedupe some deezer and spotify functionalities.
